requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
frozenlist==1.8.0
fuzzywuzzy==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.11
magic-filter==1.0.12
maxapi==0.9.15
multidict==6.7.1
propcache==0.4.1
puremagic==1.30
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
PyNaCl==1.6.2
python-telegram-bot==20.8
pytz==2025.2
requests==2.32.5
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.22.0
